requirements/common.txt
Django==1.6.5
South==0.8.4
django_pipeline==1.3.24
eco==1.1.0rc3.1
django-pipeline-eco==0.1.5
djangorestframework==2.3.13
django-recaptcha==1.0
pillow==2.4.0